requirements.txt
telethon==1.42.0
aiogram==3.24.0
python-dotenv==1.2.1
aiohttp==3.13.3
